do not pass a nil Context to logger.Log()
wrappedStmt and wrappedTx was uninitialized with a nil context in
Prepare() and Close().
This caused that in some cases logger.Log() was called with a nil ctx,
the nil ctx could then get forwarded to an external logger
implementation.
Passing a nil context causes a panic in some logger implementations.

A nil context should never be passed, as stated by the context package
documentation:

> Do not pass a nil Context, even if a function permits it. Pass
> context.TODO if you are unsure about which Context to use.
https://pkg.go.dev/context#pkg-overview

Initialize ctx with context.TODO() instead of nil, if no other ctx is
available.