Add support for DataDog
The contribution was created in whole or in part by me during the course of my work at Secureworks, Inc., and I have the right and approval to submit it under the open source license indicated by the project maintainers. I understand and agree that this project and the contribution are public and that a record of the contribution is maintained indefinitely and may be redistributed consistent with this project or the open source license(s) involved. If accepted, this contribution becomes owned by the project under control of the project maintainers.